01_Areas/Codebases/ailcc/automations/python/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-generativeai
fastapi
uvicorn
python-dotenv
requests
pydantic
browser-use
playwright
langchain-google-genai
python-dotenv
